requirements.txt
# Core AI/ML
langgraph==0.0.33
langchain-openai==0.0.8
langchain-core==0.1.33
openai==1.3.9
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
spacy==3.7.2
scikit-learn==1.3.2

# Web & API
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.0
sqlalchemy==2.0.23
alembic==1.12.1

# Deployment
docker==6.1.3
boto3==1.34.0